依存関係にserdeおよびtomlを追加し、TransmissionData構造体にシリアライズ機能を実装。送信形式をファイルに書き込む処理をmain関数に追加。

+++ src/encoder.rs
@@ -25,7 +25,7 @@ impl Encoded {
     pub fn to_transmission_data(&self) -> TransmissionData {
         let random_matrix = self.random_matrix.iter().map(|x| x.num as u8).collect();
         let data = self.value.num as u64;
-        return TransmissionData { data, random_matrix , prime: self.value.prime };
+        return TransmissionData { data, random_matrix , prime: self.value.prime as u64};
     }
 }
